requirements.txt
discord.py==1.7.3
pypresence==4.2.0
discord-py-slash-command==1.0.7
python-dotenv==0.21.0
sqlite3==2.6.0
